Added the startup html layout for our course project
Added the index.html, main.css and other files necessary for the course project. In the index.html (I have added the homepage and contents, in the Prelims section I have already started adding the topics: The Internet, The World Wide Web, The WAIS, The Gopher Protocol, The Usenet and The Hypertext Transfer Protocol (HTTP) and it's subtopics: HTTP Fundamentals, HTTP uses a request-response standard protocol, HTTP is a stateless communications protocol and HTTP provides support for other functionalities.